requirements.txt
CairoSVG==2.1.1
PyYAML==3.12
Shapely==1.5.17.post1
cairocffi==0.8.0
pyproj==1.9.5.1
tqdm==4.19.4
